Update migration versions and refactor migration class structure
- Changed migration versions for CreateUsersMigration and AlterUsersMigration to new timestamps for better versioning.
- Updated migration class names to follow CamelCase convention.
- Refactored version type from Int32 to Int64 across migration classes and related schema definitions for improved compatibility with larger version numbers.
- Adjusted database schema to reflect changes in version data type from integer to bigint.

+++ src/schemas/app_schema.cr
@@ -5,7 +5,7 @@ AppSchema = CQL::Schema.define(
   table :cql_schema_migrations do
     primary :id, Int32
     text :name
-    integer :version
+    bigint :version
     timestamps
   end
 
@@ -17,4 +17,5 @@ AppSchema = CQL::Schema.define(
     text :phone, null: true
     timestamps
   end
+
 end
